+++ src/events.ts
@@ -44,7 +44,9 @@ export type TaskEvents = 'TaskExit' | 'TaskStderr' | 'TaskStdout'
 
 export type WindowEvents = 'WinLeave' | 'WinEnter' | 'WinClosed'
 
-export type AllEvents = BufEvents | EmptyEvents | MoveEvents | TaskEvents | WindowEvents
+export type TabEvents = 'TabNew' | 'TabClosed'
+
+export type AllEvents = BufEvents | EmptyEvents | MoveEvents | TaskEvents | WindowEvents | TabEvents
   | InsertChangeEvents | 'CompleteDone' | 'TextChanged' | 'MenuPopupChanged' | 'TermClose'
   | 'InsertCharPre' | 'FileType' | 'BufWinEnter' | 'BufWinLeave' | 'VimResized' | 'TermExit'
   | 'DirChanged' | 'OptionSet' | 'Command' | 'BufReadCmd' | 'GlobalChange' | 'InputChar'
@@ -200,6 +202,7 @@ class Events {
   public on(event: MoveEvents, handler: (bufnr: number, cursor: [number, number]) => Result, thisArg?: any, disposables?: Disposable[]): Disposable
   public on(event: InsertChangeEvents, handler: (bufnr: number, info: InsertChange) => Result, thisArg?: any, disposables?: Disposable[]): Disposable
   public on(event: WindowEvents, handler: (winid: number) => Result, thisArg?: any, disposables?: Disposable[]): Disposable
+  public on(event: TabEvents, handler: (tabnr: number) => Result, thisArg?: any, disposables?: Disposable[]): Disposable
   public on(event: 'TextInsert', handler: (bufnr: number, info: InsertChange, character: string) => Result, thisArg?: any, disposables?: Disposable[]): Disposable
   public on(event: 'FloatBtnClick', handler: (bufnr: number, index: number) => Result, thisArg?: any, disposables?: Disposable[]): Disposable
   public on(event: 'TextChanged', handler: (bufnr: number, changedtick: number) => Result, thisArg?: any, disposables?: Disposable[]): Disposable